+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.0.0] - 2024-11-27
 ### Added
 - PersistenceSettings.TryDelete function
 - PersistenceSettings.TryDeleteAll function
@@ -81,7 +83,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/Persistence/compare/2.2.0...main
+[Unreleased]: https://github.com/HyagoOliveira/Persistence/compare/3.0.0...main
+[3.0.0]: https://github.com/HyagoOliveira/Persistence/tree/3.0.0/
 [2.2.0]: https://github.com/HyagoOliveira/Persistence/tree/2.2.0/
 [2.1.1]: https://github.com/HyagoOliveira/Persistence/tree/2.1.1/
 [2.1.0]: https://github.com/HyagoOliveira/Persistence/tree/2.1.0/